don't throw an error when initializing (#40)
Shouldn't be throwing an error if we cant connect to redis when initializing the class. In jupiter we call this in an initializer https://github.com/ualbertalib/jupiter/blob/05f6c0b0fa67139040f2c6e0a7126c00d0bef09b/config/initializers/oaisys.rb#L14.  When we're precompiling assets for the Docker container this error is thrown because Redis isn't started/available.

We check for connection in every other instance method and raise a ConnectionError there so we should be okay (seems like redis gem does something similar, doesn't check in initialization but checks when you actually try to do something with redis).